pluto: fix build warnings against musl c-library
Use POSIX defined locations for poll.h and sys/wait.h.
Cast cmsg_len to size_t as the struct member may be an int.

+++ programs/pluto/iface_udp.c
@@ -852,7 +852,7 @@ static bool check_msg_errqueue(const struct iface_endpoint *ifp, short interest,
 				llog(RC_LOG, logger,
 				     "unknown cmsg: level %d, type %d, len %zu",
 				     cm->cmsg_level, cm->cmsg_type,
-				     cm->cmsg_len);
+				     (size_t)cm->cmsg_len);
 			}
 		}
 	}
